Solve some issue on convert currency

+++ src/main/resources/liquibase/changesets/001_create_table.sql
@@ -2,7 +2,7 @@ create table if not exists currency_schema.currency_pairs
 (
     base_currency    varchar not null,
     convert_currency varchar not null,
-    base_amount      bigint  not null,
-    convert_amount   bigint  not null,
+    base_amount      numeric(30,6)  not null,
+    convert_amount   numeric(30,6)  not null,
     primary key (base_currency, convert_currency, base_amount)
 );
